refactor: rename to downloadworkflow
