+++ packages/cli/src/cli/commands/translate.ts
@@ -3,8 +3,8 @@ import { TranslateFlags } from '../../types/index.js';
 import { Settings } from '../../types/index.js';
 import {
   FileTranslationData,
-  executeDownloadTranslationsWorkflow,
-} from '../../workflows/downloadTranslations.js';
+  executeDownloadWorkflow,
+} from '../../workflows/download.js';
 import { createFileMapping } from '../../formats/files/fileMapping.js';
 import copyFile from '../../fs/copyFile.js';
 import flattenJsonFiles from '../../utils/flattenJsonFiles.js';
@@ -36,7 +36,7 @@ export async function handleTranslate(
       settings.defaultLocale
     );
     // Check for remaining translations
-    await executeDownloadTranslationsWorkflow({
+    await executeDownloadWorkflow({
       fileVersionData: fileVersionData,
       jobData: jobData,
       branchData: branchData,
